# tests/contracts/api-specification.yml
...
openapi: 3.0.3
info:
  title: MCP Memory Service API
  version: "6.6.1"
  description: |
    API contract for MCP Memory Service - defines actual response formats
    and status codes that clients can expect.

    CRITICAL NOTES:
    - Server returns HTTP 200 for both success and failure cases
    - Use the 'success' field in response body to determine actual result
    - All endpoints use /api prefix
servers:
  - url: https://memory.local:8443/api
    description: Default HTTPS server with self-signed certificate
  - url: http://localhost:8000/api
    description: Development HTTP server
security:
  - BearerAuth: []
paths:
...
  /memories:
    post:
      summary: Store a memory
      description: |
        Store a new memory in the service.

        CRITICAL: Always returns HTTP 200, regardless of success/failure!
        Check the 'success' field in response body to determine actual result.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  description: Memory content to store
                tags:
                  type: array
                  items:
                    type: string
                  default: []
                memory_type:
                  type: string
                  default: "note"
                metadata:
                  type: object
                  default: {}
      responses:
        '200':
          description: Request processed (check success field!)
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    title: Success
                    required:
                      - success
                      - message
                      - content_hash
                      - memory
                    properties:
                      success:
                        type: boolean
                        enum: [true]
                      message:
                        type: string
                        example: "Memory stored successfully"
                      content_hash:
                        type: string
                      memory:
                        $ref: '#/components/schemas/Memory'
                  - type: object
                    title: Duplicate
                    required:
                      - success
                      - message
                      - content_hash
                    properties:
                      success:
                        type: boolean
                        enum: [false]
                      message:
                        type: string
                        example: "Duplicate content detected"
                      content_hash:
                        type: string
                      memory:
                        type: 'null'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Invalid API key"
  /search:
    get:
      summary: Search memories by content
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: n_results
          in: query
          schema:
            type: integer
            default: 5
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        memory:
                          $ref: '#/components/schemas/Memory'
                        relevance_score:
                          type: number
                          minimum: 0
                          maximum: 1
  /memories/search/tags:
    get:
      summary: Search memories by tags
      parameters:
        - name: tags
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated list of tags
      responses:
        '200':
          description: Tag search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  memories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Memory'
  /memories/{content_hash}:
    delete:
      summary: Delete a memory by content hash
      parameters:
        - name: content_hash
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deletion result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '404':
          description: Memory not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key for authentication
  schemas:
    Memory:
      type: object
      required:
        - content
        - content_hash
        - tags
        - memory_type
        - created_at_iso
      properties:
        content:
          type: string
        content_hash:
          type: string
        tags:
          type: array
          items:
            type: string
        memory_type:
          type: string
        metadata:
          type: object
        created_at:
          type: number
        created_at_iso:
          type: string
          format: date-time
        updated_at:
          type: number
        updated_at_iso:
          type: string
          format: date-time

# Contract Test Cases
...